Fixing Configuration issues
- Changed configurations that should be converted to milliseconds.
- Implemented Day to Millisecond converter.
- Added missing JavaDocs.
- Added relevant unit tests.

+++ src/main/java/com/l2jserver/gameserver/config/BufferServiceConfiguration.java
@@ -49,7 +49,7 @@ public interface BufferServiceConfiguration extends Reloadable {
 	
 	@Key("HealCooldown")
 	@ConverterClass(Seconds2MillisecondsConverter.class)
-	int getHealCooldown();
+	long getHealCooldown();
 	
 	@Key("MaxUniqueLists")
 	int getMaxUniqueLists();
